run docker processes as non-root
Because we're spinning up multiple containers and some of them need to share
something, I've made it so all the processes run with the same UID and GID of
host user spinning up the services. To communicate current UID and GID to the
container startup scripts, udco, a wrapper around docker-compose, has been created.