remove nested dependencies from requirements.txt
Dependencies-of-dependencies are not exactly required in requirements.txt,
and they make removing direct dependencies that are not required anymore harder.
All indirect dependencies are commented out in case we need them again later, and
are left for pip to figure out.

PyJWT was excluded from requirements.txt despite it being imported in account/views.py.
This was added.