bump up dependencies
Most importantly, Django is bumped from 2.2 to 3.0, because bigger number better number.
This might break stuff, but a bit of manual testing seems to show that things work, so
there's that.

Turns out djangorestframework-jwt is currently unmaintained, and there is a maintained
fork(drf-jwt). BUT, due to some obscure reason, I've failed to get it working with this
project; I've kinda tracked the root reason(ie. view's post method and stuff like that are
meant to be invoked as a part of django's dispatch system, and not otherwise since the dispatch
system does some initialization work before calling them), but I didn't write the auth system
and I fear fixing that may cause some other problem. Thus, no upgrade to drf-jwt. This is
desirable, but not a priority 'absolutely needed' thing.