add user creation hook
Now UserStorage and root directory is created at user creation,
not at first upload. This allows us to assume they 'just exist', and
limit DB locking to that particular user's row. Also removed code
that tried to create UserStorage object if it doesn't exist, because
now if it doesn't exist it's an hard error.