overhaul directory structure
All objects that can logically be a inside a directory (directory, file, upload-in-progress)
now inherit from the new added DirectoryEntry, which allows us to enforce unique name with in
a directory for all of them, and also cuts some repeating fields in the models.

The kind field is considered experimental and could be removed if it doesn't quite work.

Note that the migration added for this change has some bullshit in it: in particular, it adds
some randomly generated uuid as 'parent' of everything that's not root. Good thing we're still indev.